+++ RenderCore/Render/RHI/D3D12/RHIRenderingItem_D3D12.h
@@ -18,7 +18,7 @@ public:
 	virtual void UploadInstanceDataToBuffer(IRHIContext* Context)override;
 	virtual void MarkInstanceDataDirty(std::vector<FColor>& Colors, UINT Width, UINT Height)override;
 
-	virtual std::shared_ptr<IRHIIndexBuffer> BuildIndexBuffer(std::vector<uint16_t>& InIndices)override;
+	virtual std::shared_ptr<IRHIIndexBuffer> BuildIndexBuffer(std::vector<uint32_t>& InIndices)override;
 	virtual std::shared_ptr<IRHIVertexBuffer> BuildVertexBuffer(std::vector<FVertex>& InVertices)override;
 	virtual std::shared_ptr<IRHIVertexBuffer> BuildVertexBuffer(std::vector<FSkinVertex>& InVertices)override;
